Add Locomotion callbacks + Fix memory corruption on unload (#8)
Allows developers to customize `CBaseNPC_Locomotion`'s behavior by giving the option to override its default functions through callbacks, in similar fashion to `NextBotActionFactory`. Developers can now customize how the bot climbs, jumps, collides with other entities, and even determine if an entity is traversable or not. Developers can also still call the default behavior if desired in their callbacks using the `CallBaseFunction` native. The default behavior will be called if no callback is set.

`CallBaseFunction` is restricted to be used within a callback function only as it's supposed to be a private thing. Outside of callbacks, developers should be calling the functions virtually anyway using the `ILocomotion`/`NextBotGroundLocomotion` natives.

The `CBaseNPC.GetLocomotion` native has been changed to return a `CBaseNPC_Locomotion` object. Backwards compatibility with older plugins won't be an issue as `CBaseNPC_Locomotion` inherits from `NextBotGroundLocomotion`.

This also removes the `SetCollisionBounds` native, as it's been deprecated and is no longer defined in the .inc files.

Co-authored-by: Kenzzer <kenzzer@users.noreply.github.com>