Fix gamedata based entity factory (#37)
Due to the nature of member function ptrs on linux, the extension would auto offset the class object ptr allocated by the game. Leading to misaligned memory.

On the topic of misaligned memory, we weren't really guarding against it. So just in the event the engine changes how memory is allocated, or that plugin provides an object size that doesn't allow for 4 bytes aligned memory, we fix that before allocating anything.